% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{calcR0}
\alias{calcR0}
\title{caluclate R0}
\usage{
calcR0(beta = 10, gamma1 = 1/4, hs = 0.26, hr = 12.95)
}
\arguments{
\item{beta}{transmission rate}

\item{gamma1}{rate of I->R}

\item{hs}{shape of gamma-distributed one-time partnership rate}

\item{hr}{rate param of gamma-distributed one-time partnership rate}
}
\description{
caluclate R0
}
