% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_compare}
\alias{model_compare}
\title{Compare observed and modelled data for the basic model}
\usage{
model_compare(state, observed, pars)
}
\arguments{
\item{state}{State vector for the end of the current day. This is
assumed to be filtered following \code{\link[=model_index]{model_index()}} so contains
rows corresponding to new infections and scaled time.}

\item{observed}{Observed data. This will be a list with elements
\code{Ytravel}, \code{Yendog} and \code{Yunk}.}

\item{pars}{A list of parameters}
}
\value{
A vector of log likelihoods, the same length as the number
of particles (the number of columns in the modelled state)
}
\description{
Compare observed and modelled data from the basic model. This
conforms to the mcstate interface.
}
