% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_index}
\alias{model_index}
\title{Index of model variables}
\usage{
model_index(info)
}
\arguments{
\item{info}{The result of running the \verb{$info()} method on an
initialised \link{model}}
}
\value{
A list with element \code{run}, indicating the locations of the
compartments used in \link{model_compare}
}
\description{
Index of "interesting" elements for the model. This function
conforms to the mcstate interface.
}
