% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{pgf_h}
\alias{pgf_h}
\title{PGF of one-time partnership network}
\usage{
pgf_h(x, hs, hr)
}
\arguments{
\item{x}{model output}

\item{hs}{shape of gamma-distributed one-time partnership rate}

\item{hr}{rate param of gamma-distributed one-time partnership rate}
}
\description{
PGF of one-time partnership network
}
