% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_filter}
\alias{model_filter}
\title{Create particle filter}
\usage{
model_filter(data, ..., use_compiled_compare = FALSE)
}
\arguments{
\item{data}{Data to use}

\item{...}{Additional arguments to \link[mcstate:particle_filter]{mcstate::particle_filter}'s constructor}

\item{use_compiled_compare}{Logical, indicating if we should use
the compiled compare function rather than \link{model_compare}. This
will likely be faster if several threads are used.}
}
\description{
Create a particle filter; just a convenience function around some
mcstate defaults.
}
